refactor: refactor name widget banner item
