+++ lib/presenter/widgets/home/banner_item.dart
